Added RTC
STM8's internal RTC has been activated. The master-controller (e.g. Thread module) can now transmit a date and time into the STM8 controller and this will automatically update the time and count it in realtime.
Also a disconnection will be shown on the display in case the Thread router loses power